fix: Symlinked working dir not mounted within container.
Snakemake currenlty only mounts the real path (i.e. path without
symlinks) into containers. This can lead to problems in case
scripts refer to the symlinked-version of the path (#2837,
possibly also #2959)

This commit detects when the PWD has symlinks, and then
additionally also mounts the symlinked version of the PWD
into the container.